fix the deadloop issue and support for custom timeout
testDone is scheduled every 500ms, but the termination
check is counter === number, so, there is high possibility
that after every 500ms, the counter is bigger than number,
then testDone will go to deadloop.

Another issue is socket.io does not support to specify
a custom timeout value, the default value is 20000ms.